Refactor analyzeItem into AddressAnalyzer
