+++ packages/backend/src/core/discovery/analysis/getCodeLink.ts
@@ -1,0 +1,12 @@
+import { EthereumAddress } from '@l2beat/shared'
+
+export function getCodeLink(
+  address: EthereumAddress,
+  implementations: EthereumAddress[] | undefined,
+) {
+  const addresses = [address]
+  if (implementations) {
+    addresses.push(...implementations)
+  }
+  return `https://etherscan.deth.net/address/${addresses.join(',')}`
+}
