--- packages/shared/src/services/AddressAnalyzer.test.ts
@@ -1,61 +1,0 @@
-import { expect, mockObject } from 'earl'
-import { providers } from 'ethers'
-
-import { EthereumAddress } from '../types'
-import { AddressAnalyzer } from './AddressAnalyzer'
-import { EtherscanClient } from './etherscan'
-
-describe(AddressAnalyzer.name, () => {
-  it('can detect an EOA', async () => {
-    const provider = mockObject<providers.Provider>({
-      getCode: async () => '0x',
-    })
-    const etherscanClient = mockObject<EtherscanClient>({
-      getContractSource: async () => ({} as any),
-    })
-    const addressAnalyzer = new AddressAnalyzer(provider, etherscanClient)
-    const result = await addressAnalyzer.analyze(
-      EthereumAddress('0x11223344556677889900aabbccddeeff11223344'),
-    )
-    expect(result).toEqual({ type: 'EOA', name: '<EOA 11223344>' })
-  })
-
-  it('can detect an unverified contract', async () => {
-    const provider = mockObject<providers.Provider>({
-      getCode: async () => '0x112233',
-    })
-    const etherscanClient = mockObject<EtherscanClient>({
-      getContractSource: async () =>
-        ({ ABI: 'Contract source code not verified' } as any),
-    })
-    const addressAnalyzer = new AddressAnalyzer(provider, etherscanClient)
-    const result = await addressAnalyzer.analyze(
-      EthereumAddress('0x11223344556677889900aabbccddeeff11223344'),
-    )
-    expect(result).toEqual({
-      type: 'Contract',
-      verified: false,
-      name: '<Unverified 11223344>',
-    })
-  })
-
-  it('can detect a verified contract', async () => {
-    const provider = mockObject<providers.Provider>({
-      getCode: async () => '0x112233',
-    })
-    const etherscanClient = mockObject<EtherscanClient>({
-      getContractSource: async () =>
-        ({ ABI: '[{"type":"string"}]', ContractName: 'Foo' } as any),
-    })
-    const addressAnalyzer = new AddressAnalyzer(provider, etherscanClient)
-    const result = await addressAnalyzer.analyze(
-      EthereumAddress('0x11223344556677889900aabbccddeeff11223344'),
-    )
-    expect(result).toEqual({
-      type: 'Contract',
-      verified: true,
-      name: 'Foo',
-      abi: [{ type: 'string' }],
-    })
-  })
-})
