+++ packages/shared/src/services/index.ts
@@ -1,4 +1,3 @@
-export * from './AddressAnalyzer'
 export * from './coingecko'
 export * from './etherscan'
 export * from './HttpClient'
